update Vendor to use Resources
